blog: First show_attached_content blog_entry
Give the blog model a GenericForeignKey, to be able to attach foreign
object to a blog entry. Each type needs to have it's own template set
in show_attached_content. As of now, it is possible to attach an event,
but the templatetag needs refining and implementation for music etc.